finished room + building cards
