+++ app/screens/Buildings/building_list.tsx
@@ -6,10 +6,11 @@ import {
 	SafeAreaView,
 	ScrollView,
 	Pressable, 
-	Button,
 } from "react-native";
 import type { BuildingStackScreenProps  } from '../types';
 import { FreeRoomsAPIContext } from '../../contexts';
+import BuildingCard from '../../components/BuildingCard';
+
 
 import Ionicons from '@expo/vector-icons/Ionicons';
 
@@ -18,7 +19,6 @@ export default function BuildingList({ route, navigation } : BuildingStackScreen
 	const handleFilterPress = () => {
 		navigation.navigate("Building Filter");
 	}
-	
 	
 	const { buildings, rooms, onRefresh } = useContext(FreeRoomsAPIContext);
 	
@@ -37,47 +37,14 @@ export default function BuildingList({ route, navigation } : BuildingStackScreen
 		<SafeAreaView style={ styles.container }>
 			<ScrollView style={ styles.scrollView }>
 				<View style={ styles.container }>
-					<Text style={ styles.main_heading }>Buildings at UNSW</Text>
 					<View style={ styles.textContainer }>
-						<Text style={{ paddingBottom: 10, }}>
-							Scroll and click on a Building Card!
-						</Text>
 					</View>
-					{ buildings.map((building, index) => <Card key={index} nav={navigation} {...building} /> )}
+					{ buildings.map((building, index) => <BuildingCard key={index} nav={navigation} rooms={rooms} {...building} /> )}
 				</View>
 			</ScrollView>
 		</SafeAreaView>
 		
 	)
-}
-
-function Card({ nav, name, id }) {
-	
-	const handlePress = (name: string, id : string ) => {
-		nav.navigate("Building", { buildingName: name , buildingId : id });
-	}
-	
-	return (
-		
-			<View style={{
-					width: '100%',
-					height: 100,
-					alignItems: 'flex-start',
-					paddingHorizontal: 10,
-					paddingVertical: 20,
-					borderWidth: 1,
-					borderRadius: 10,
-					borderColor: 'lightgray',
-					backgroundColor: 'orange',
-					marginVertical: 10,
-			}}>
-				<Pressable onPress={() => { handlePress(name, id) }} style={styles.pressableCard}>
-					<Text style={styles.subHeading}>
-						{ `${name} (${id})`}
-					</Text>
-				</Pressable>
-			</View>
-	);
 }
 
 const styles = StyleSheet.create({
@@ -89,7 +56,6 @@ const styles = StyleSheet.create({
 
   },
   scrollView: {
-    paddingTop: 20,
     paddingHorizontal: 10,
 		width: '100%',
   },
@@ -98,23 +64,8 @@ const styles = StyleSheet.create({
     fontWeight: 'bold',
     paddingBottom: 20,
   },
-  subHeading: {
-    fontSize: 15,
-    fontWeight: 'bold',
-    paddingBottom: 10,
-    width: '100%',
-
-  },
-  textBody: {
-    fontSize: 10,
-  },
   textContainer: {
     flex: 1, 
     padding: 10,
-  },
-  pressableCard: {
-    flex: 1,
-    justifyContent: 'center',
-    width: '100%',
   }
 });
